chore: move gateway list utils
